Renamed connection file..
Added support for connecting to prod DB. To do this, updated mongoose to version >6

This update doesn't need @types/mongoose anymore so the dependency was removed. Connection code was updated to work without @types/mongoose and with mongoose version > 6
