+++ src/helper/Constants.ts
@@ -3,6 +3,8 @@ export default class Constants {
 	static readonly YOUTUBE_API_KEY = process.env.YOUTUBE_API_KEY
 	static readonly HEART_API_KEY = process.env.HEART_API_KEY
 
+	static readonly HEART_API_DB_BASE_URI = process.env.HEART_API_DB_BASE_URI
+
 	static readonly SKC_CHANNEL_ID = 'UCBZ_1wWyLQI3SV9IgLbyiNQ'
 	static readonly RBF_CHANNEL_ID = 'UCKSWgpqi8BnNgDYh_jwYQNw'
